apollo-smith: generate document from apollo-parser(#192)
This commit introduces two improvements: allows `apollo-smith` to generate a document built on `apollo-parser` with a `parser-impl` feature flag, and introduces a few semantic validations to the generation process.

Semantic validations currently include:
- Directives used in the document must already be defined
- Directives must be unique in a given Directive Location
- Default values must be of correct type
- Input values must be of correct type
- All type extensions are applied to an existing type
- Field arguments in fragments and operation definitions must be defined on original type and must be of correct type

+++ crates/apollo-parser/src/parser/grammar/value.rs
@@ -192,7 +192,7 @@ enum Test @dir__one(int_value: -10) {
                         if let ast::Value::IntValue(val) =
                             argument.value().expect("Cannot get argument value.")
                         {
-                            let i: i64 = val.into();
+                            let i: i32 = val.into();
                             assert_eq!(i, -10);
                         }
                     }
